requirements.txt
certifi==2020.6.20
chardet==3.0.4
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyserial==3.4
requests==2.24.0
urllib3==1.25.11
zipp==3.4.0
